test(test_ps): convert to pytest
